ClientRequester: Keep track of latest success/failure Date
This marks the beginning of branch issue-6384-part1.
Its goal shall be to fix part of what is demanded at
https://bugs.freenetproject.org/view.php?id=6384

Specifically, the branch shall implement two UI columns for
downloads and uploads: "Latest success" and "Latest failure".
They shall show the Dates which are added in this commit.
What the branch will NOT implement is keeping track of failure reasons:
As far as I understand, class ClientRequester is the "root" class which
stores downloads/uploads, so it should be the place which stores such
information. But it doesn't actually receive reasons for success/failure
in its callbacks. Thus, it would be more work to implement than I can
currently afford to spend on this.

As for this specific commit, the following is notable:
There already *is* storage for "last activity" in ClientRequest (and
another class whose name I don't remember right now), but IMHO that
class is more low level and thus storage should be in the class to which
this commit adds it. The branch will deal with propagating the
information down to all classes which need it so it can be displayed in
the UI.