SplitfileProgressEvent: Consume Date latestSuccess & latestFailure
The previous commit added Dates latestSuccess & latestFailure to class
ClientRequester. That class provides an abstract function
innerNotifyClients() whose implementations have the job of creating
SplitfileProgressEvent objects which contain its state.

The following commits shall adapt the innerNotifyClients()
implementations to pass the new dates to the SplitfileProgressEvents.
For being able to do this, this commit amends SplitfileProgressEvent to
provide storage for the dates, and also to consume them in the
constructors.

Notice: getDescription() is not adapted to include it since as far as I
can see, this function is only used for output in simulations and the
text-mode client interface. So its a debug feature mostly.