*RequestStatus: Consume Date latestSuccess & latestFailure
Commit 86a5a06ec9ac380b12d9ba780871938bca370def added Dates
latestSuccess & latestFailure to class ClientRequester.

Commit f64b6ee9f42a97adc034f48a33db7c93029ab143 added the same to
SplitfileProgressEvent.

Commit 481026107800c3c32765bbdd69e23a1fb45918f1 adapted ClientRequester
implementations to propagate the dates to generated
SplitfileProgressEvent.

As far as I understand class RequestStatus which this commit deals with,
it serves the following purpose related to ClientRequester /
SplitfileProgressEvent: Class RequestStatusCache provides a cached copy
of the latest state of a request as it was propagated by
SplitfileProgressEvent. This is cached instead of being queried directly
from ClientRequester for concurrency issues, i.e. not having to wait for
locks.
RequestStatus is what is actually used by the user interface to query
the information.
Thus, RequestStatus needs to be changed to also be able to store
latestSuccess & latestFailure, which this commit does.

The callers of the RequestStatus constructors in child classes are also
updated to pass through the dates.
The following commit will deal with adapting non-child classes to the
new constructors.