ClientRequest: Remove obsolete, now unused lastActivity tracking
The previous commits moved the tracking to ClientRequester already, and
this code is unused now.

Notice:
In a recent IRC discussion (Freenode IRC channel #freenet, 2015-02-21),
toad said that the changes which I'm doing as part of this branch should
be accompanied by incrementing the serialVersionUID of all changed
Serializable classes.

He said that the download/upload queue should handle this safely in a
way which causn't loss of downloads/uploads: It has a backup of the base
data to restart them, and the backup is stored in a
non-Java-serialization fileformat. Incrementing the serialVersionUID
should cause it to discard the Java-serialization data and use that
backup.